Adding explicit import of read_dot/modifying code
Implementing a workaround to a networkx bug that was introduced after this was published.  See https://github.com/networkx/networkx/issues/1984 for more information.  Along with explicitly importing read_dot, you need to install pydotplus.  Unsure if pydot is still required as I didn't test on a host without it.