Add support for showing only the shortest cycles
When analysing dependency issues and a cycle exists, then this cycle
will show up in a lot of different forms. Removing just the shortest
cycle will also remove all the longer ones. Therefore, showing only
the shortest ones significantly reduces the output and thereby the
amount of data to analyse.